+++ Pixelwave/Classes/Support/Parsers/TextureAtlasParser/TextureAtlasParsers/PXTPAtlasParser.m
@@ -197,7 +197,7 @@
 		if (trimmed)
 		{
 			// Calculates the padding values (Also converts to points).
-			short *padding = cFrame->padding;
+			float *padding = cFrame->padding;
 			// Top
 			padding[0] = (spriteSourceSize.origin.y) * invScaleFactor;
 			// Right
